requirements.txt
PyQt6~=6.9.1
sounddevice~=0.5.2
webrtcvad==2.0.10
pytesseract~=0.3.13
opencv-python==4.9.0.80
numpy~=2.2.6
pillow~=10.2.0
requests~=2.31.0
python-dotenv~=1.0.1
Pygments~=2.17.2
faster-whisper~=1.1.1
PyQt6>=6.0
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.31.0
pydub>=0.25.1
SpeechRecognition>=3.10.1
whisper @ git+https://github.com/openai/whisper.git